Time: 2 ms (95.7%), Space: 48.6 MB (23.77%) - LeetHub

+++ 0167-two-sum-ii-input-array-is-sorted/0167-two-sum-ii-input-array-is-sorted.java
@@ -1,6 +1,6 @@
 class Solution {
     public int[] twoSum(int[] numbers, int target) {
-        short l=0;
+        int l=0;
         int r=numbers.length-1;
         while(l<r){            
             if(numbers[l]+numbers[r]==target) return new int []{l+1,r+1};
